% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{data_trans}
\alias{data_trans}
\title{Transform data for estimation}
\usage{
data_trans(y, model = NULL, log.vars = NULL, diff.vars = NULL,
  freq = NULL, detect.diff = F, detect.growth = F, level = 0.01,
  panelID = NULL, timeID = NULL)
}
\arguments{
\item{y}{Multivariate time series of data values. May also be a data frame containing a date column}

\item{model}{Estimated model from ms_dcf_estim}

\item{log.vars}{Character vector of variables to be logged}

\item{diff.vars}{Character vector of unit root variables to be differenced}

\item{freq}{Seasonality of the data}

\item{detect.diff}{Logical, detect which variables to difference}

\item{detect.growth}{Logical, detect which variables to log}

\item{level}{Significance level of statistical tests (0.01, 0.05, 0.1)}

\item{panelID}{Column name that identifies the cross section of the data}

\item{timeID}{Column name that identifies the date}

\item{detect.formula}{Logical, detect lag length of the dynamic common factor to include in each observation equation using the cross correlation function up to a max of 3}
}
\value{
list containing differenced data (yy_d) and standardized data (yy_s)
}
\description{
Transform data for estimation
}
