% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{detect_frequency}
\alias{detect_frequency}
\title{Detect data frequency}
\usage{
detect_frequency(y, timeID)
}
\arguments{
\item{y}{Multivariate time series of data values. May also be a data frame containing a date column}

\item{timeID}{Column name that identifies the date}
}
\description{
Detect data frequency
}
