% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{ms_dcf_filter}
\alias{ms_dcf_filter}
\title{Kim filter (Hamilton + Kalman filter) an estimated model from ms_dcf_estim}
\usage{
ms_dcf_filter(y, model, plot = F)
}
\arguments{
\item{y}{Multivariate time series of data values. May also contain a date column.}

\item{model}{Structural time series model estimated using ms_dcf_estim.}

\item{plot}{Logial, whether to plot the output or not.}
}
\value{
List of data tables containing the filtered and smoothed series.
}
\description{
Kim filter (Hamilton + Kalman filter) an estimated model from ms_dcf_estim
}
\examples{
ms_dcf_filter(y = DT[, c("date", "y")], model = model)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
