% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{set_constraints}
\alias{set_constraints}
\title{Set the constraints for estimation}
\usage{
set_constraints(yy_s, theta, n_states, panelID, timeID)
}
\arguments{
\item{yy_s}{Multivariate time series of standardized data values from data_trans}

\item{theta}{Named ector of initial coefficient values}

\item{n_states}{Number of states to include in the Markov switching model}

\item{panelID}{Column name that identifies the cross section of the data}

\item{timeID}{Column name that identifies the date}
}
\value{
vector of initial coefficient values
}
\description{
Set the constraints for estimation
}
\examples{
set_constraintsy(yy_s = yy_s, theta = theta, n_states = n_states, panelID = panelID, timeID = timeID)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
