% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{SSmodel_ms}
\alias{SSmodel_ms}
\title{State space model for markov switching mean
Creates a state space model in list form
yt = At + Ht * Bt + e_t
Bt = D_t + Ft * B_tl + u_t}
\usage{
SSmodel_ms(par, yt, n_states, ms_var, panelID = NULL, timeID = NULL,
  init = NULL)
}
\arguments{
\item{par}{Vector of named parameter values}

\item{yt}{Multivariate time series of data values}

\item{panelID}{Column name that identifies the cross section of the data}

\item{timeID}{Column name that identifies the date}
}
\value{
List of space space matrices
}
\description{
State space model for markov switching mean
Creates a state space model in list form
yt = At + Ht * Bt + e_t
Bt = D_t + Ft * B_tl + u_t
}
\examples{
SSmodel_ms(par, yt, panelID, timeID)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
